backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-google-genai
pymupdf
pytesseract
Pillow
beautifulsoup4
requests
faiss-cpu
langchain
langchain_community